FInal product
The subroutine is not implemented, so it is not functioning properly. It works well when we call it inside a label since it has the RET opcode at the end of the label so it is somehow mimicking a subroutine but when the instructions are executed sequentially, an issue arises. This occurs because we have utilized the label by placing a return statement at the end of the label. Consequently, this leads to a stack overflow when the return(RET) statement is encountered during a direct sequential execution.